+++ contracts/orderbook/OrderBookIntegrity.sol
@@ -7,8 +7,8 @@ import "./OrderBook.sol";
 import "../type/LibCast.sol";
 import "../idempotent/LibIdempotentFlag.sol";
 
-uint constant FLAG_INDEX_CLEARED_ORDER = 0;
-uint constant FLAG_INDEX_CLEARED_COUNTERPARTY = 1;
+uint256 constant FLAG_INDEX_CLEARED_ORDER = 0;
+uint256 constant FLAG_INDEX_CLEARED_COUNTERPARTY = 1;
 
 contract OrderBookIntegrity is StandardIntegrity {
     using LibCast for function(uint256) pure returns (uint256)[];
@@ -20,7 +20,11 @@ contract OrderBookIntegrity is StandardIntegrity {
         Operand,
         StackTop stackTop_
     ) internal pure returns (StackTop) {
-        integrityState_.scratch = IdempotentFlag.unwrap(IdempotentFlag.wrap(integrityState_.scratch).set(FLAG_INDEX_CLEARED_ORDER));
+        integrityState_.scratch = IdempotentFlag.unwrap(
+            IdempotentFlag.wrap(integrityState_.scratch).set(
+                FLAG_INDEX_CLEARED_ORDER
+            )
+        );
         function(uint256) internal view returns (uint256) fn_;
         return integrityState_.applyFn(stackTop_, fn_);
     }
@@ -30,7 +34,11 @@ contract OrderBookIntegrity is StandardIntegrity {
         Operand,
         StackTop stackTop_
     ) internal pure returns (StackTop) {
-        integrityState_.scratch = IdempotentFlag.unwrap(IdempotentFlag.wrap(integrityState_.scratch).set(FLAG_INDEX_CLEARED_COUNTERPARTY));
+        integrityState_.scratch = IdempotentFlag.unwrap(
+            IdempotentFlag.wrap(integrityState_.scratch).set(
+                FLAG_INDEX_CLEARED_COUNTERPARTY
+            )
+        );
         function(uint256, uint256) internal view returns (uint256) fn_;
         return integrityState_.applyFn(stackTop_, fn_);
     }
